+++ servers/dsub/jobs/test/test_jobs_controller_local.py
@@ -18,7 +18,6 @@ PROCESS_NOT_FOUND_MESSAGE = 'Process not found yet'
 
 class TestJobsControllerLocal(BaseTestCases.JobsControllerTestCase):
     """ JobsController integration tests for local provider """
-
     @classmethod
     def setUpClass(cls):
         super(TestJobsControllerLocal, cls).setUpClass()